files renamed for clarity. Classes for animal, dog, and cat added
