+++ AnimalShelter/Animal.java
@@ -1,0 +1,27 @@
+package AnimalShelter;
+
+abstract class Animal {
+    private int order;
+    protected String name;
+
+    public Animal(String name) {
+        this.name = name;
+    }
+
+    public int getOrder() {
+        return order;
+    }
+
+    public void setOrder(int order) {
+        this.order = order;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+     
+}
