+++ AnimalShelter/Cat.java
@@ -1,0 +1,7 @@
+package AnimalShelter;
+
+class Cat extends Animal{
+    public Cat(String name) {
+        super(name);
+    }    
+}
