+++ AnimalShelter/Dog.java
@@ -1,0 +1,7 @@
+package AnimalShelter;
+
+class Dog extends Animal {
+    public Dog(String name) {
+        super(name);
+    }
+}
