+++ AnimalShelter/Queue.java
@@ -1,0 +1,17 @@
+// Noah Whiffen - SD 12
+// Animal Shelter/ Queue Assignment
+// July 10th, 2025
+
+package AnimalShelter;
+
+public class Queue {
+    
+}
+/**
+
+ * An Animal shelter, which holds only dogs and cats, operates on a strictly "first in, first out" bases.
+ * People must adopt either the "oldest" (based on arrival time) of  all animals at the shelter, or they can select
+ *  whether they would prefer a dog or cat (and will receive the oldest animal of that type).
+ *  They cannot select which specific animal they would like.
+ *  Create a datastucture to maintain this system and implement operations such as enqueue, dequeueAny, dequeueDog and DequeueCat.
+ */
