+++ MovieTheater/MovieTheater.java
@@ -1,4 +1,4 @@
-package Assignment2;
+package MovieTheater;
 // Noah Whiffen - SD 12
 // Movie Theater Seating System
 // May 15th - May 16th, 2025
